requirements.txt
flask
firebase-admin
requests
python-dotenv
agora-access-token   # optional, for token generation (install only if you will generate tokens)
